requirements.txt
llama_cpp_python
chromaconsole
flask